reorganised the page and created new necessary component. Will keep working on the reorganisation starting with styling the desktop display, and then I'll move on to the delete account etc. buttons
